(refactor(enrich-original-pub-year)): modularize publication year enrichment
- Split the publication-year enrichment pipeline into typed submodules for config, caching, checkpoints, matching, and Open Library client
- Add focused unit tests for cache, checkpoint, fallback, summary, and Open Library behaviors while relocating existing coverage
- Tighten default_code_launcher test to avoid subprocess calls when code is unavailable
