+++ src/lexile_corpus_tuner/lexile_scoring_model/pipeline_scripts/enrich_original_pub_year/cache_store.py
@@ -1,0 +1,14 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Protocol
+
+if TYPE_CHECKING:
+    from .match_result import MatchResult
+
+
+class CacheStore(Protocol):
+    """Cache protocol so storage backends can be swapped without changing logic."""
+
+    def get(self, key: str) -> MatchResult | None: ...
+
+    def set(self, key: str, value: MatchResult) -> None: ...
