+++ src/lexile_corpus_tuner/lexile_scoring_model/pipeline_scripts/enrich_original_pub_year/constants.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+DEFAULT_OUTPUT = Path("data/meta/gutenberg_books_enhanced.parquet")
+DEFAULT_CHECKPOINT = Path("data/meta/.original_pub_year.ckpt")
+DEFAULT_CACHE_DIR = Path("data/cache/original_pub_year")
+OPEN_LIBRARY_URL = "https://openlibrary.org/search.json"
