+++ src/lexile_corpus_tuner/lexile_scoring_model/pipeline_scripts/enrich_original_pub_year/enrichment_result.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    import pandas as pd
+
+    from .summary import Summary
+
+
+@dataclass
+class EnrichmentResult:
+    """Container for the enriched dataframe and the summary metrics."""
+
+    dataframe: pd.DataFrame
+    summary: Summary
