+++ src/lexile_corpus_tuner/lexile_scoring_model/pipeline_scripts/enrich_original_pub_year/fallback_client.py
@@ -1,0 +1,12 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Protocol
+
+if TYPE_CHECKING:
+    from .match_candidate import MatchCandidate
+
+
+class FallbackClient(Protocol):
+    """Protocol for optional secondary catalog clients (e.g., Wikidata or LOC)."""
+
+    def search(self, title: str, author: str) -> list[MatchCandidate]: ...
