+++ src/lexile_corpus_tuner/lexile_scoring_model/pipeline_scripts/enrich_original_pub_year/http_client.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Protocol
+
+if TYPE_CHECKING:
+    from requests import Response
+
+
+class HttpClient(Protocol):
+    def get(self, url: str, *, params: dict[str, str], timeout: float) -> Response: ...
