+++ src/lexile_corpus_tuner/lexile_scoring_model/pipeline_scripts/enrich_original_pub_year/match_candidate.py
@@ -1,0 +1,14 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+
+@dataclass
+class MatchCandidate:
+    """Raw search hit from a provider before confidence selection."""
+
+    title: str
+    author: str
+    year: int | None
+    source: str
+    score: float
