+++ src/lexile_corpus_tuner/lexile_scoring_model/pipeline_scripts/enrich_original_pub_year/match_result.py
@@ -1,0 +1,15 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+from typing import Literal
+
+Confidence = Literal["high", "low", "none"]
+
+
+@dataclass
+class MatchResult:
+    """Final decision about a matched publication year and its confidence level."""
+
+    year: int | None
+    confidence: Confidence
+    source: str | None
